Initial take on announced subscription interest.
Instead of every aircraft publishing their coordinates to every topic
and potentially flooding every map_client with data that has to be
filtered every 10s, we instead move this assignment to the aircraft.

By informing the Flight controller with information in the form of a
topic encoded with lat_lng data representing a box, it then becomes
available to every aircraft to get a list of available topics to
broadcast to.

The intention is then that every aircraft that is withing a box then
broadcast its information to every box it is inside.

This also has the potential of becoming an extreme task - as this has
the potential to be just as many unique boxes as there are connected
users. This could perhaps be solved with less resolution on the lat_lng
data.